% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/forecast.lassovar.R
\name{forecast.lassovar}
\alias{forecast.lassovar}
\title{Multiple forecasts a VAR using the Lasso or adaptive lasso.}
\usage{
forecast.lassovar(dat, ntrain, horizon = 1, fc.window = c("fix",
  "expanding"), fc.type = c("recursive", "direct"), ic = c("BIC", "AIC"),
  mc = FALSE, silent = FALSE, ...)
}
\arguments{
\item{dat}{A data.frame.}

\item{ntrain}{The number of training observations to use.}

\item{horizon}{The forecast horizon, default to 1.}

\item{fc.window}{Should the number of training observations be fixed or expanding when the sample increases?}

\item{fc.type}{Should the forecasts be recursive or direct?}

\item{ic}{Information criterion used to select the penalty parameter: BIC or AIC.}

\item{mc}{Parallelize the forecasts with the multicore package. If you want to parallelize the equation-by-equation esitmation of the lasso var, set mc=FALSE and mclas=TRUE. In both cases, use the optional parameter ncores to select the numbers of cores to use.}

\item{silent}{Should output be printed?}

\item{...}{Optional options...}
}
\value{
A list.
}
\description{
This function is designed to compute pseudo out-of-sample forecasting experiments. For out-of-sample forecasts use the predict method. The function will produce h-step ahead forecasts either directly, by fitting a h-step ahead VAR, or recursively. The model is trained on ntrain observations, and as many h-step ahead forecasts as possible are computed on the remainder of the sample.
}

