\name{lassovar-package}
\alias{lassovar-package}
\docType{package}
\title{Estimated Vector Autoregression (VAR) using the Lasso or some of its variants}
\description{This package aims at providing a flexible way to estimate VARs using the Lasso or the adaptive Lasso using lagged values of the dependent variables or common factors extracted from the dependent variables.}
\details{
\tabular{ll}{
Package: \tab lassovar\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-01\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Laurent Callot (lcallot@vu.nl)
Maintainer: Laurent Callot (lcallot@vu.nl) 
}
