% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_prediction_data.R
\name{prepare_prediction_data}
\alias{prepare_prediction_data}
\title{prepare_prediction_data}
\usage{
prepare_prediction_data(object, fc.data)
}
\arguments{
\item{object}{lassovar estimation object}

\item{fc.data}{model data.frame}
}
\value{
data matrix augmented with lag columns
}
\description{
prepare_prediction_data
}
