% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.lassovar.R
\name{residuals.lassovar}
\alias{residuals.lassovar}
\alias{residuals}
\alias{resid}
\title{The residuals method of class lassovar.}
\usage{
\method{residuals}{lassovar}(object, ...)
}
\arguments{
\item{object}{A lassovar object.}

\item{...}{use short=TRUE for a shorter printout.}
}
\description{
The residuals method of class lassovar.
}
