% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lassovar.R
\name{lassovar}
\alias{lassovar}
\title{Estimates a Vector Autoregressive model by mean of Lasso or adaptive Lasso}
\usage{
lassovar(dat, exo = NULL, lags = 1, ic = c("BIC", "AIC"),
  adaptive = c("none", "ols", "lasso", "group", "ridge"), post = FALSE,
  mc = FALSE, ncores = NULL, dfmax = NULL, ...)
}
\arguments{
\item{dat}{a data.frame containing the series.}

\item{exo}{Optional, a data frame of same length as dat containing exogenous variables.}

\item{lags}{the number of consecutive lags in the model.}

\item{ic}{Optional, the information criterion to use for selecting the penalty parameter. Default to BIC.}

\item{adaptive}{Optional, initial estimator for the adaptive Lasso. Can be chosen between: none (default, plain Lasso), ols, lasso, ridge.}

\item{post}{Optional, Should a post Lasso OLS be estimated, default FALSE.}

\item{mc}{Optional, default to FALSE. Should the equation-by-equation estimation be parallelized using mclapply from the parallel package?}

\item{ncores}{Optional, the number of cores to use in case of parallelization.}

\item{dfmax}{Optional, the maximum number of variables in the model excluding the intercept. An option of glmnet, it exits the algorithm when the penalty is small enough that more than dmax variables are included in the model. Incresease the speed tremendously for large VARs.}

\item{...}{Optional options...}
}
\value{
A lassovar object.
}
\description{
Estimates a Vector Autoregressive model by mean of Lasso or adaptive Lasso
}
\examples{
\dontrun{
dat <- data.frame(matrix(rnorm(100),ncol=5))
lv.mod <- lassovar(dat,lags=1)
lv.mod.ada <- lassovar(dat,lags=1,adaptive ='ols')
}
}

