% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.lassovar.R
\name{predict.lassovar}
\alias{predict.lassovar}
\title{predict.lassovar}
\usage{
\method{predict}{lassovar}(object, fc.data, ...)
}
\arguments{
\item{object}{A lassovar object.}

\item{fc.data}{A data frame with the new data.}

\item{...}{unused}
}
\description{
predict.lassovar
}

