% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.lassovar.R
\name{summary.lassovar}
\alias{summary.lassovar}
\title{The summary method of class lassovar.}
\usage{
\method{summary}{lassovar}(object, ...)
}
\arguments{
\item{object}{A lassovar object.}

\item{...}{use short=TRUE for a shorter printout.}
}
\description{
The summary method of class lassovar.
}

